apiprj/requirements.txt
# Python packages required for main.py

django
djangorestframework

psycopg2

requests
drf-yasg

gunicorn